fix(Map): Refactor map settings and prepared it to easier using
